+++ test/e2e/wallet/wallet-join-api.test.ts
@@ -19,10 +19,13 @@ describe('/wallets/{walletId}/join endpoint', () => {
   let database: Sequelize;
   let server: FastifyInstance;
   let expirationTime: number;
+  let enableSync: boolean;
   beforeAll(async () => {
     database = await setupDatabase(false);
     server = setupServer(database);
-    expirationTime = parseEnvironment().EXPIRATION_TIME;
+    const environment = parseEnvironment();
+    enableSync = environment.ENABLE_SYNC;
+    expirationTime = environment.EXPIRATION_TIME;
   });
 
   afterAll(async () => {
@@ -30,7 +33,7 @@ describe('/wallets/{walletId}/join endpoint', () => {
   });
 
   beforeEach(async () => {
-    await database.sync({ force: true });
+    if (enableSync) await database.sync({ force: true });
   });
 
   test('should return the wallet state after join', async () => {
